Login Fix
There were 2 recent changes that prevented the login from working.

We had thought that project_name in the call signiture was not used, however it seems
important for function name resolution.

The automatic logout seems to be logging out as soon as I am logging in.  This should
be based on the timeout of the token, so a separate api should be opened up on atmosphere
to see if the token has timedout.